feature : paypoint와 member의 point Integer로 수정

+++ backend/member/src/main/java/com/paymong/member/member/dto/response/ModifyPointResDto.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ModifyPointResDto {
-    Long point;
+    Integer point;
 }
