+++ backend/member/src/main/java/com/paymong/member/paypoint/dto/AddPointReqDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class AddPointReqDto {
     String content;
-    Long point;
+    Integer point;
 }
